Refactor AtlassianClient into modular client architecture
- Create BaseAtlassianClient with OAuth and HTTP functionality
- Extract JiraClient, ConfluenceClient, ServiceDeskClient from monolithic AtlassianClient
- Update modules to use their respective client classes
- Streamline server.py to focus on initialization and tool registration
- Maintain 9.99/10 pylint score with clean separation of concerns

This resolves the too-many-methods issue by distributing client methods
across specialized client classes while maintaining the same functionality.
